+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Chart Type :
+- Chart Bar (Simple, Horizontal, Vertical & Custom)
+- Chart Line (Simple, Point, Area & Custom)
+- Chart Pie (Simple, Donut & Custom)
+- Chart Gauge (Simple & Custom)